bot_app/requirements.txt
aioredis
fastapi
uvicorn
httpx
requests
python-dotenv
langchain
openai
chromadb
soundfile
colorlog
aiofile